Fix type errors

+++ lib/Doctrine/ORM/Tools/EntityRepositoryGenerator.php
@@ -80,7 +80,7 @@ class <className> extends <repositoryName>
      */
     private function getClassNamespace($fullClassName)
     {
-        $namespace = substr($fullClassName, 0, strrpos($fullClassName, '\\'));
+        $namespace = substr($fullClassName, 0, (int) strrpos($fullClassName, '\\'));
 
         return $namespace;
     }
